Make --update update the cache a single item at a time, locking the 	database only around the modification of each item.  Add a new option, 	--rebuild, which completely rebuilds the cache from scratch, locking 	the database for the entire duration of the rebuild.
	* src/openvasmd.c (main): Add --rebuild option.

	* src/ompd.c (init_ompd): Update param doc.
	(write_to_scanner): Add update states by refering to ompd_nvt_cache_mode.
	(serve_omp): Set ompd_nvt_cache_mode directly from client_socket, which
	determines which update mode to use.  Add update states to select setup.

	* src/omp.c (init_omp_process): Update param doc.

	* src/tasks_sql.h (init_manage_process): Update param doc.  Only remove
	NVT and preference caches when rebuilding the cache.  Move COMMIT to
	manage_complete_nvt_cache_update.
	(make_nvt_from_nvti): Add argument to control removal of NVT before
	inserting.
	(manage_complete_nvt_cache_update): Add mode argument; COMMIT if rebuild.

	* src/manage.h: Update headers.

	* src/otp.h (scanner_init_state_t): Add new states.

	* src/otp.c (parse_scanner_preference_value): Add update state.
	(parse_scanner_plugin_list_tags): Pass NVT removal flag.
	(process_otp_scanner_input): Add update states.  In particular, pass mode
	to manage_complete_nvt_cache_update.

	* README: Update usage message.

git-svn-id: svn+ssh://scm.wald.intevation.org/openvas/trunk/openvas-manager@5971 423fd1db-d629-0410-8442-d21db03e70f4

+++ src/tasks_sql.h
@@ -2028,7 +2028,8 @@ next_task (task_iterator_t* iterator, task_t* task)
  *
  * Open the SQL database.
  *
- * @param[in]  update_nvt_cache  If true, clear the NVT cache.
+ * @param[in]  update_nvt_cache  0 operate normally, -1 just update NVT cache,
+ *                               -2 just rebuild NVT cache.
  * @param[in]  database          Location of manage database.
  */
 void
@@ -2039,13 +2040,12 @@ init_manage_process (int update_nvt_cache, const gchar *database)
 
   if (task_db)
     {
-      if (update_nvt_cache)
+      if (update_nvt_cache == -2)
         {
           sql ("BEGIN EXCLUSIVE;");
           sql ("DELETE FROM nvts;");
           sql ("DELETE FROM nvt_preferences;");
           sql ("DELETE FROM meta WHERE name = 'nvts_checksum';");
-          sql ("COMMIT;");
         }
       return;
     }
@@ -2077,11 +2077,13 @@ init_manage_process (int update_nvt_cache, const gchar *database)
 
   if (update_nvt_cache)
     {
-      sql ("BEGIN EXCLUSIVE;");
-      sql ("DELETE FROM nvts;");
-      sql ("DELETE FROM nvt_preferences;");
-      sql ("DELETE FROM meta WHERE name = 'nvts_checksum';");
-      sql ("COMMIT;");
+      if (update_nvt_cache == -2)
+        {
+          sql ("BEGIN EXCLUSIVE;");
+          sql ("DELETE FROM nvts;");
+          sql ("DELETE FROM nvt_preferences;");
+          sql ("DELETE FROM meta WHERE name = 'nvts_checksum';");
+        }
     }
   else
     {
@@ -6269,12 +6271,13 @@ nvt_family (nvt_t nvt)
 /**
  * @brief Make an nvt from an nvti.
  *
- * @param[in]  nvti  NVTI.
+ * @param[in]  nvti    NVTI.
+ * @param[in]  remove  Whether to remove the NVT from the cache first.
  *
  * @return An NVT.
  */
 nvt_t
-make_nvt_from_nvti (const nvti_t *nvti)
+make_nvt_from_nvti (const nvti_t *nvti, int remove)
 {
   /** @todo Freeing string literals. */
   gchar *quoted_version, *quoted_name, *quoted_summary, *quoted_description;
@@ -6299,6 +6302,12 @@ make_nvt_from_nvti (const nvti_t *nvti)
                                    : "");
   quoted_family = sql_quote (nvti_family (nvti) ? nvti_family (nvti) : "");
 
+  if (remove)
+    {
+      sql ("BEGIN EXCLUSIVE;");
+      sql ("DELETE FROM nvts WHERE oid = '%s';", nvti_oid (nvti));
+    }
+
   sql ("INSERT into nvts (oid, version, name, summary, description, copyright,"
        " cve, bid, xref, tag, sign_key_ids, category, family)"
        " VALUES ('%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s', '%s',"
@@ -6316,6 +6325,9 @@ make_nvt_from_nvti (const nvti_t *nvti)
        quoted_sign_key_ids,
        nvti_category (nvti),
        quoted_family);
+
+  if (remove)
+    sql ("COMMIT;");
 
   g_free (quoted_version);
   g_free (quoted_name);
@@ -6506,11 +6518,14 @@ update_config_caches ()
 
 /**
  * @brief Complete an update of the NVT cache.
+ *
+ * @param[in]  mode  -1 updating, -2 rebuilding.
  */
 void
-manage_complete_nvt_cache_update ()
+manage_complete_nvt_cache_update (int mode)
 {
   update_config_caches ();
+  if (mode == -2) sql ("COMMIT;");
 }
